refactor: rename "variants" to "editors", move view only canvas to components
