+++ src/components/questionStructure/Task/variants/handwritingTasks/components/ViewOnlyCanvas.tsx
