+++ src/components/questionStructure/Task/variants/handwritingTasks/components/viewOnlyCanvas.scss
