+++ src/components/questionStructure/Task/variants/handwritingTasks/editors/MathsProcessedHandwritingEditor.tsx
