+++ src/components/questionStructure/Task/variants/handwritingTasks/editors/RawHandwritingEditor.tsx
